Move remoteconfig tests (#7318)
